module d'installation - part 1.2

+++ flore/home/install.php
@@ -91,8 +91,9 @@ switch ($action)
 /*Formulaire d'installation------------------------------------------------------------*/
 default :
 case "install-param":	{
-
-	/*test installation*/
+	
+	/*--------------------------------------*/
+	/*Renseignement des valeurs du formulaire*/
 	if (file_exists("../../_INCLUDE/install-ok.txt"))
 		{
 		require_once ("../../_INCLUDE/config_sql.inc.php");		
@@ -117,33 +118,38 @@ case "install-param":	{
 			}
 		}
 	else
+		{
+		require_once ("../../_INCLUDE/config_sql.inc.example.php");	
 		foreach ($rub  as $key => $val)	{$rub_ok[$key] = 'f'; $desc[$key] = "";}
-			
-	
+		}
+
 	echo ("<div id=\"fiche\" >");
 	echo ("<form method=\"POST\" id=\"form1\" class=\"form1\" name=\"edit\" action=\"\" >");
 	echo ("<center><input type=\"hidden\" name=\"action\" value=\"install-set\" />");
+	if (!file_exists("../../_INCLUDE/install-ok.txt"))
+	{
 	//------------------------------------------------------------------------------ EDIT LR GRP1
 			echo ("<div id=\"radio1\">");    
 			echo ("<fieldset style=\"width: 50%;\"><LEGEND>Connexion au serveur de base de données</LEGEND>");
 					echo ("<table border=0 width=\"100%\"><tr valign=top >");
 					echo ("<td style=\"width: 800px;\">");
-						metaform_text ("Hôte","",50,"","host","localhost");
-						metaform_text ("Port","",25,"","port","5432");
-						metaform_text ("Utilisateur admin","",50,"","user","postgres");
-						metaform_pw ("Mot de passe admin","",50,"","mdp","");
+						metaform_text ("Hôte","",50,"","host",SQL_server);
+						metaform_text ("Port","",25,"","port",5432);
+						metaform_text ("Utilisateur admin","",50,"","user",SQL_admin_user);
+						metaform_pw ("Mot de passe admin","",50,"","mdp",SQL_admin_pass);
 					echo ("</td></tr></table>");	
 				echo ("</fieldset>");
 	//------------------------------------------------------------------------------ EDIT LR GRP2 
 			echo ("<fieldset  style=\"width: 50%;\"><LEGEND> Création de la base de données </LEGEND>");
 					echo ("<table border=0 width=\"100%\"><tr valign=top >");
 					echo ("<td style=\"width: 800px;\">");
-						metaform_text ("Nom de la base","",50,"","dbname","codex");
-						metaform_text ("Utilisateur codex","",50,"","user_codex","admin_codex");
-						metaform_pw ("Mot de passe codex","",20,"","mdp_codex","admin_codex");
+						metaform_text ("Nom de la base","",50,"","dbname",SQL_base);
+						metaform_text ("Utilisateur codex","",50,"","user_codex",SQL_user);
+						metaform_pw ("Mot de passe codex","",20,"","mdp_codex",SQL_pass);
 					echo ("</td></tr></table>");	
 				echo ("</fieldset>");
-	//------------------------------------------------------------------------------ EDIT LR GRP2 
+	}
+	//------------------------------------------------------------------------------ EDIT LR GRP3
 			echo ("<fieldset style=\"width: 50%;\"><LEGEND> Choix des rubriques à installer </LEGEND>");
 					echo ("<table border=0 width=\"100%\"><tr valign=top >");
 					echo ("<td style=\"width: 800px;\">");
@@ -167,79 +173,95 @@ case "install-param":	{
 case "install-set":	{
 	echo ("<div id=\"fiche\" >");
 	if (isset($_POST["host"]) AND isset($_POST["port"]) AND isset($_POST["user"]) AND isset($_POST["mdp"]) AND isset($_POST["dbname"]) AND isset($_POST["user_codex"]) AND isset($_POST["mdp_codex"]))
+		{$host = $_POST["host"];$port = $_POST["port"];$user = $_POST["user"];$mdp = $_POST["mdp"];$dbname = $_POST["dbname"];$user_codex = $_POST["user_codex"];$mdp_codex = $_POST["mdp_codex"];}
+	elseif (file_exists("../../_INCLUDE/install-ok.txt"))
+		{require_once ("../../_INCLUDE/config_sql.inc.php");	$host = SQL_server;$port = SQL_port;$user = SQL_user;$mdp = SQL_pass;$dbname = SQL_base;}
+	else
+		echo ("Problème de connexion<button id=\"return-button\">retour au menu</button></center>");
+	
+	$conn_admin = connexion ($host,$port,$user,$mdp,"postgres");
+
+	if ($conn_admin)
 		{
-		$host = $_POST["host"];$port = $_POST["port"];$user = $_POST["user"];$mdp = $_POST["mdp"];$dbname = $_POST["dbname"];$user_codex = $_POST["user_codex"];$mdp_codex = $_POST["mdp_codex"];
-		$conn_admin = connexion ($host,$port,$user,$mdp,"postgres");
-		if ($conn_admin)
+		echo ("connexion établie<BR>"); 
+		
+		/*------------------*/
+		/*Création de la BDD*/
+		$conn_codex = connexion ($host,$port,$user,$mdp,$dbname);
+		if (!$conn_codex)
 			{
-			echo ("connexion établie<BR>"); 
-			
-			/*------------------*/
-			/*Création de la BDD*/
-			$conn_codex = connexion ($host,$port,$user,$mdp,$dbname);
-			if (!$conn_codex)
-				{
-				$result = pg_query($conn_admin,"CREATE DATABASE $dbname ENCODING = 'UTF8' LC_COLLATE = 'French_France.1252' LC_CTYPE = 'French_France.1252';");
-				echo ("La base de données $dbname a été créée<BR>"); 
-				}
-			else
-				echo ("La base de données $dbname existait déjà<BR>"); 
-			
-			/*-------------------------------*/
-			/*Création de l'utilisateur CODEX*/
-			$conn_codex = connexion ($host,$port,$user,$mdp,$dbname);
-			if (!pg_query($conn_codex,"SELECT 1 FROM pg_roles WHERE rolname='$user_codex';"))		
-				{
-				$result = pg_query($conn_codex,"CREATE USER $user_codex PASSWORD $mdp_codex;");
-				echo ("L'utilisateur $user_codex a été créé<BR>"); 
-				}
-			else
-				echo ("L'utilisateur $user_codex existait déjà<BR>"); 
-			
-			/*-------------------*/	
-			/*Structure de la BDD*/
-			foreach ($rub as $key => $val)
-				{
-				if ($_POST[$key] == 't')
-					{
-					$sql = "../../_DATA/bdd_codex_archi_$key.sql";
-					$query = create_schema($sql,$user_codex);
-					$result = pg_query($conn_codex,$query);
-					echo ("L'architecture de la $valeur a été implémentée<BR>"); 
-					}
-				else
-					echo ("L'architecture de la $valeur existait déjà ou n'a pas été selectionnée<BR>"); 
-				}
-			
-			
-			/*----------------------------*/
-			/*Intégration des référentiels*/
-			$result = pg_query($conn_codex,"SELECT cd_ref FROM refnat.taxrefv80_utf8 LIMIT 1");
-			$row = pg_fetch_row($result);			
-			if (empty($row))
-				{
-				// $query = create_data($user_codex);
-				// $result = pg_query($conn_codex,$query);
-				echo ("Les référentiels de la base de données $dbname ont été implémentés<BR>"); 
-				}
-			else
-				echo ("Les référentiels de la base de données $dbname avaient déjà été implémentés<BR>"); 
-			
-			/*Bouton finalisation de l'install*/
-			echo ("<form method=\"POST\" id=\"form1\" class=\"form1\" name=\"edit\" action=\"\" >");			
-			echo ("<center><input type=\"hidden\" name=\"action\" value=\"install-finish\" />");
-			echo ("<button id=\"install-finish-button\">Fin de l'installation</button></center>");
-			echo ("</form>");
+			$result = pg_query($conn_admin,"CREATE DATABASE $dbname ENCODING = 'UTF8' LC_COLLATE = 'French_France.1252' LC_CTYPE = 'French_France.1252';");
+			echo ("La base de données $dbname a été créée<BR>"); 
 			}
 		else
+			echo ("La base de données $dbname existait déjà<BR>"); 
+		
+		/*-------------------------------*/
+		/*Création de l'utilisateur CODEX*/
+		$conn_codex = connexion ($host,$port,$user,$mdp,$dbname);
+		if (!pg_query($conn_codex,"SELECT 1 FROM pg_roles WHERE rolname='$user_codex';"))		
 			{
-			echo ("Problèmes de connexion<BR>");
-			echo ("<button id=\"return-button\">retour au menu</button></center>");
+			$result = pg_query($conn_codex,"CREATE USER $user_codex PASSWORD $mdp_codex;");
+			echo ("L'utilisateur $user_codex a été créé<BR>"); 
 			}
+		else
+			echo ("L'utilisateur $user_codex existait déjà<BR>"); 
+		
+		/*-------------------*/	
+		/*Structure de la BDD*/
+		foreach ($rub as $key => $val)
+			{
+			if ($_POST[$key] == 't')
+				{
+				$sql = "../../_DATA/bdd_codex_archi_$key.sql";
+				$query = create_schema($sql,$user_codex);
+				// $result = pg_query($conn_codex,$query);
+				echo ("L'architecture de la $valeur a été implémentée<BR>"); 
+				}
+			else
+				echo ("L'architecture de la $valeur existait déjà ou n'a pas été selectionnée<BR>"); 
+			}
+		
+		
+		/*----------------------------*/
+		/*Intégration des référentiels*/
+		$result = pg_query($conn_codex,"SELECT cd_ref FROM refnat.taxrefv80_utf8 LIMIT 1");
+		$row = pg_fetch_row($result);			
+		if (empty($row))
+			{
+			// $query = create_data($user_codex);
+			// $result = pg_query($conn_codex,$query);
+			echo ("Les référentiels de la base de données $dbname ont été implémentés<BR>"); 
+			}
+		else
+			echo ("Les référentiels de la base de données $dbname avaient déjà été implémentés<BR>"); 
+			
+			
+		/*------------------------------------------*/
+		/*parametrage du ficher de conf sql_connect*/
+		if (!file_exists("../../config_sql.inc.php"))
+			{
+			copy ("../../config_sql.inc.example.php","../../config_sql.inc.php");
+			$sql_file = file_get_contents("../../config_sql.inc.php");
+			$sql_file = str_replace("localhost",$host,$sql_file);
+			$sql_file = str_replace("5432",$port,$sql_file);
+			$sql_file = str_replace("user_codex",$user_codex,$sql_file);
+			$sql_file = str_replace("codex_user",$mdp_codex,$sql_file);
+			$sql_file = str_replace("postgres",$user,$sql_file);
+			$sql_file = str_replace("admin_pass",$mdp,$sql_file);
+			$sql_file = str_replace("codex",$dbname,$sql_file);
+			}
+		
+		
+		/*Bouton finalisation de l'install*/
+		echo ("<form method=\"POST\" id=\"form1\" class=\"form1\" name=\"edit\" action=\"\" >");			
+		echo ("<center><input type=\"hidden\" name=\"action\" value=\"install-finish\" />");
+		echo ("<button id=\"install-finish-button\">Fin de l'installation</button></center>");
+		echo ("</form>");
 		}
 	else
 		{
-		echo ("Tous les paramètres requis ne sont pas remplis");
+		echo ("Problèmes de connexion<BR>");
 		echo ("<button id=\"return-button\">retour au menu</button></center>");
 		}
 	echo ("</div>");
